# swagger-cartao-v1.yaml
openapi: 3.0.3
info:
  title: vasco-bank-cartao - OpenAPI 3.0
  description: ''
  version: 1.0.11

servers:
  - url: https:/vasco-bank-cartao/api/v1
  
tags:
  - name: cartaoCredito
    description: Operations about user
paths:

  /cartaoCredito/{contaId}/all:
    get:
      tags:
        - cartaoCredito
      summary: Retorna informações de todos cartoes de credito de uma conta caso existir
      description: This can only be done by the logged in user.
      operationId: cartaoall
      parameters:
        - name: contaId
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cartoes'
            application/xml:
              schema:
                $ref: '#/components/schemas/cartoes'
  /cartaoCredito/{contaId}/{cartaoId}:
    get:
      tags:
        - cartaoCredito
      summary: Retorna informação do cartao de credito de uma conta caso existir
      description: This can only be done by the logged in user.
      operationId: cartao
      parameters:
        - name: contaId
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
        - name: cartaoId
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cartaoCredito'
            application/xml:
              schema:
                $ref: '#/components/schemas/cartaoCredito'
    delete:
      tags:
        - cartaoCredito
      summary: Deletar cartao do usuario de conta especificada
      description: Creates list of users with given input array
      operationId: cartaoDelete
      parameters:
        - name: contaId
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
        - name: cartaoId
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cartaoCredito'          
            application/xml:
              schema:
                $ref: '#/components/schemas/cartaoCredito'
        default:
          description: successful operation
  /cartaoCredito/{contaId}:
    post:
      tags:
        - cartaoCredito
      summary: Criar cartao na conta
      description: Creates list of users with given input array
      operationId: cartaoPost
      parameters:
        - name: contaId
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/cartaoCredito'
          application/xml:
            schema:
              $ref: '#/components/schemas/cartaoCredito'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/cartaoCredito'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cartaoCredito'          
            application/xml:
              schema:
                $ref: '#/components/schemas/cartaoCredito'
        default:
          description: successful operation

  /cartaoCredito/{contaId}/{cartaoId}/limite:
    get:
      tags:
        - cartaoCredito
      summary: Recuperar limite do cartao
      description: ''
      operationId: cartaoCreditoLimite
      parameters:
        - name: contaId
          in: path
          description: ID da conta do usuario
          required: true
          schema:
            type: integer
            format: int64
        - name: cartaoId
          in: path
          description: ID do cartao
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/limite'          
            application/xml:
              schema:
                $ref: '#/components/schemas/limite'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
          
  /cartaoCredito/{contaId}/{cartaoId}/fatura/:
    get:
      tags:
        - cartaoCredito
      summary: Recuperar fatura do cartao
      description: ''
      operationId: cartaoCreditoFatura
      parameters:
        - name: contaId
          in: path
          description: ID da conta do usuario
          required: true
          schema:
            type: integer
            format: int64
        - name: cartaoId
          in: path
          description: ID do cartao
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fatura'          
            application/xml:
              schema:
                $ref: '#/components/schemas/fatura'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found

  /cartaoCredito/{contaId}/{cartaoId}/extrato/{intervalo}:
    get:
      tags:
        - cartaoCredito
      summary: Recuperar extrato de acordo com um intervalo
      description: ''
      operationId: cartaoCreditoExtrato
      parameters:
        - name: contaId
          in: path
          description: ID da conta do usuario
          required: true
          schema:
            type: integer
            format: int64
        - name: cartaoId
          in: path
          description: ID do cartao do cartao
          required: true
          schema:
            type: integer
            format: int64
        - name: intervalo
          in: path
          description: Enum de identificacao do intervalo do extrato
          required: true
          schema:
              type: string
              enum:
              - DIA
              - MES
              - ANUAL
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/extrato'          
            application/xml:
              schema:
                $ref: '#/components/schemas/extrato'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found     
          
  /cartaoCredito/{contaId}/{cartaoId}/pagamento:
    post:
      tags:
        - cartaoCredito
      summary: Realiza pagamento do cartao do usuario
      description: ''
      operationId: cartaoCreditoPagamento
      parameters:
        - name: cartaoId
          in: path
          description: ID do cartao do usuario
          required: true
          schema:
            type: integer
            format: int64
        - name: contaId
          in: path
          description: ID da conta do usuario
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/limite'
          application/xml:
            schema:
              $ref: '#/components/schemas/limite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/limite'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/limite'          
            application/xml:
              schema:
                $ref: '#/components/schemas/limite'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found   
    
components:
  schemas:
    cartoes:
          type: array
          xml:
            wrapped: true
            name: cartoes
          items:
            $ref: '#/components/schemas/cartaoCredito'
         
    cartaoCredito:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        nome:
          type: string
          example: 'John K. pascal'
        numero:
          type: string
          example: '1234 1232 1234 1232'
        dataVencimento:
          type: string  
        cvv:
          type: string
          example: '123'
        limite:
          type: number
          example: 9.999
        eligivelAumentoLimite:
          type: boolean
          example: 'false'
        virtual:
          type: boolean
          example: 'false'
      xml:
        name: cartaoCredito
    limite:
        type: object
        properties:
          valor:
            type: number
            format: int64
            example: 99.999
        xml:
          name: limite
    fatura:
        type: object
        properties:
          id:
            type: integer
            format: int64
            example: 10
          valorTotal:
            type: number
            format: int64
            example: 99.999
          paga:
            type: boolean
            example: false
          dataVencimento:
            type: string
            format: date
          compras:
            type: array
            xml:
              wrapped: true
              name: compra
            items:
              $ref: '#/components/schemas/compra'
        xml:
          name: fatura
          
    compra:
        type: object
        properties:
          id:
            type: integer
            format: int64
            example: 10
          valor:
            type: number
            format: int64
            example: 99.999
          dataCompra:
            type: string
            format: date
          local:
            type: string
            example: 'uber'
          parcelas:
            type: integer
            example: 12
        xml:
          name: compra
          
    extrato:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        dataOperacao:
          type: string
          format: date
        operacao:
          type: string
          enum:
            - PIX
            - BOLETO
            - INTRABANCO
            - PAGCARTAO
      xml:
        name: extrato
          
          
           
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://vasco-bank-cartao/api/v1/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header